requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
plotly>=5.22
fpdf2>=2.7
XlsxWriter>=3.2
